start overhauling RaidCalc objects
